Created context, added all the functions and states there that I would need to use in components below. Solved problem about saving cookies and setting user. Now user can do crud operations for todos when authenticated.

+++ server/src/routes/todo.js
@@ -1,25 +1,29 @@
 const express = require('express');
+const passport = require('passport');
 const router = express.Router();
 const pool = require('../db/db');
 
 // Routes
 router.post('/todos', async (req, res) => {
   try {
-    const { description, user_id } = req.body;
-    // description will go to VALUES
-    // $1 is to add dynamic data
-    // Returning returns newly added value
-    const newTodo = await pool.query(
-      // insert into todo tables description
-      'INSERT INTO todo (description, user_id) VALUES($1, $2) RETURNING *',
-      [description, user_id]
-    );
+    if (req.isAuthenticated()) {
+      const { user_id } = req.user;
+      const { description } = req.body;
+      // description will go to VALUES
+      // $1 is to add dynamic data
+      // Returning returns newly added value
+      const newTodo = await pool.query(
+        // insert into todo tables description
+        'INSERT INTO todo (description, user_id) VALUES($1, $2) RETURNING *',
+        [description, user_id]
+      );
 
-    if (!newTodo) {
-      return res.send({ Error: 'Could not create to do' });
+      if (!newTodo) {
+        return res.send({ Error: 'Could not create to do' });
+      }
+
+      return res.json(newTodo.rows[0]);
     }
-
-    return res.json(newTodo.rows[0]);
   } catch (error) {
     return res.send(error);
   }
@@ -27,9 +31,14 @@ router.post('/todos', async (req, res) => {
 
 router.get('/todos', async (req, res) => {
   try {
-    const allTodos = await pool.query('SELECT * FROM todo');
+    if (req.isAuthenticated()) {
+      const allTodos = await pool.query(
+        'SELECT * FROM todo WHERE user_id = $1',
+        [req.user.user_id]
+      );
 
-    return res.json(allTodos.rows);
+      return res.json(allTodos.rows);
+    }
   } catch (error) {
     return res.send(error);
   }
@@ -66,13 +75,14 @@ router.put('/todos/:id', async (req, res) => {
 
 router.delete('/todos/:id', async (req, res) => {
   try {
-    const { id } = req.params;
-    const deleteTodo = await pool.query(
-      'DELETE FROM todo WHERE todo_id = $1 RETURNING *',
-      [id]
-    );
-
-    return res.send(deleteTodo.rows);
+    if (req.isAuthenticated()) {
+      const { id } = req.params;
+      const deleteTodo = await pool.query(
+        'DELETE FROM todo WHERE todo_id = $1 RETURNING *',
+        [id]
+      );
+      return res.send(deleteTodo.rows);
+    }
   } catch (error) {
     return res.send(error);
   }
